Fix float16 attention caches before SciPy sparse construction
Promote each selected channel to float64 at CSR, diagonal COO and dense construction. Keep stored NPZ arrays, coordinates and scoring equations unchanged.

Add 20 dtype regression cases and include them in run_all.sh. On SciPy 1.17.0 the new tests reproduce 8 failures before this fix and all 20 pass afterwards.

+++ experiments/unsupervised_token_graph/channels.py
@@ -84,11 +84,14 @@ def iter_channels(record, layers=None, heads=None):
                 first = (layer * n_heads + head) * count
                 ptr = pointer[first:first + count + 1]
                 begin, end = ptr[0], ptr[-1]
-                matrix = sparse.csr_matrix((values[begin:end], columns[begin:end], ptr - begin), shape=(count, tokens))
+                # Promote only this channel before construction; SciPy sparse rejects float16.
+                matrix = sparse.csr_matrix((values[begin:end], columns[begin:end], ptr - begin),
+                                           shape=(count, tokens), dtype=np.float64)
                 matrix = matrix + sparse.csr_matrix((diagonal[layer, head, p:],
-                                                    (np.arange(count), queries)), shape=(count, tokens))
+                                                    (np.arange(count), queries)),
+                                                   shape=(count, tokens), dtype=np.float64)
             else:
-                matrix = sparse.csr_matrix(attention[layer, head])
+                matrix = sparse.csr_matrix(attention[layer, head], dtype=np.float64)
             # Prompt queries are not modelled internally; the last prompt query
             # can still supply the first next-token prediction when it was saved.
             keep = np.asarray(queries) >= p - 1
